+++ src/main/kotlin/io/ionic/libs/iongeolocationlib/model/IONGLOCLocationResult.kt
@@ -11,5 +11,9 @@ data class IONGLOCLocationResult(
     val altitudeAccuracy: Float? = null,
     val heading: Float,
     val speed: Float,
-    val timestamp: Long
+    val timestamp: Long,
+    val magneticHeading: Float? = null,
+    val trueHeading: Float? = null,
+    val headingAccuracy: Float? = null,
+    val course: Float? = null
 )